food_save_backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
channels[daphne]>=4.0.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
django-cors-headers>=4.3.0
uvicorn>=0.23.0
